requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0   # optional; install if you plan to use the xgboost model
arch>=5.3.0      # required for GARCH(1,1) implementation
pyyaml>=6.0
pyarrow>=9.0.0   # for parquet read/write
tqdm>=4.65.0     # optional (progress bars)
